// src/app/models/empleado/createEmpleado.dto.ts
export class createEmpleadoDto {
    id?: number = 0;
    codigo_empleado: string = '';
    idTipoEmpleado: number = 0;
    nombre_Empleado: string = '';
    telefono: string = '';
    region_origen: string = '';
    acompanantes: number = 0;
    observaciones: string = '';
    idUsuario: number = 0;
}